Moved WooCommerce class to separate file
Finally capitalised that C baby!

Oh yeah, and I've probably broken a couple things too...
